+++ packages/css-syntax-patches-for-csstree/scripts/generate-webref-sets.mjs
@@ -112,6 +112,16 @@ export async function generate_webref_sets() {
 			property.syntax = 'rect( <top> , <right> , <bottom> , <left> ) | rect( <top> <right> <bottom> <left> ) | auto';
 		}
 
+		// https://github.com/csstree/csstree/pull/354
+		if ((
+			property.name === 'overflow-x' ||
+			property.name === 'overflow-y' ||
+			property.name === 'overflow-inline' ||
+			property.name === 'overflow-block'
+		) && property.syntax === 'visible | hidden | clip | scroll | auto') {
+			property.syntax = '| <-non-standard-overflow>';
+		}
+
 		if (!property.syntax) {
 			continue;
 		}